+++ src/sensors/hardware.rs
@@ -9,7 +9,7 @@ use esp_hal::{
 pub(super) struct SensorHardware<'a> {
     pub(super) adc1: Adc<'a, ADC1<'a>, Blocking>,
     pub(super) moisture_pin: AdcPin<GPIO2<'a>, ADC1<'a>, AdcCalCurve<ADC1<'a>>>,
-    pub(super) waterlevel_pin: AdcPin<GPIO3<'a>, ADC1<'a>, ()>,
+    pub(super) waterlevel_pin: AdcPin<GPIO3<'a>, ADC1<'a>, AdcCalCurve<ADC1<'a>>>,
     pub(super) battery_pin: AdcPin<GPIO4<'a>, ADC1<'a>, AdcCalLine<ADC1<'a>>>,
     pub(super) moisture_power_pin: Output<'a>,
     pub(super) water_level_power_pin: Output<'a>,
@@ -37,7 +37,8 @@ pub(super) async fn initialize_hardware(p: SensorPeripherals) -> SensorHardware<
     let mut adc1_config = AdcConfig::new();
     let moisture_pin = adc1_config
         .enable_pin_with_cal::<_, AdcCalCurve<ADC1>>(p.moisture_analog_pin, Attenuation::_11dB);
-    let waterlevel_pin = adc1_config.enable_pin(p.water_level_analog_pin, Attenuation::_11dB);
+    let waterlevel_pin = adc1_config
+        .enable_pin_with_cal::<_, AdcCalCurve<ADC1>>(p.water_level_analog_pin, Attenuation::_11dB);
     let battery_pin = adc1_config.enable_pin_with_cal(p.battery_pin, Attenuation::_11dB);
     let adc1 = Adc::new(p.adc1, adc1_config);